app/notifications/views.py
from rest_framework import generics
from rest_framework.permissions import IsAuthenticated
from core.models import Notification
from core.serializers import NotificationSerializer

class NotificationListView(generics.ListAPIView):
    serializer_class = NotificationSerializer
    permission_classes = [IsAuthenticated]

    def get_queryset(self):
        return Notification.objects.filter(user=self.request.user).order_by('-timestamp')