requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
channels>=3.0,<4.0
channels_redis>=3.2.0,<4.0
